Fix some post-merge issues
